INT-2040 changed signature of MessageStore.getMessageCount() to return lo

+++ spring-integration-core/src/main/java/org/springframework/integration/store/SimpleMessageStore.java
@@ -76,7 +76,7 @@ public class SimpleMessageStore extends AbstractMessageGroupStore implements Mes
 	}
 
 	@ManagedAttribute
-	public int getMessageCount() {
+	public long getMessageCount() {
 		return idToMessage.size();
 	}
